+++ CHANGELOG.md
@@ -5,9 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project
 adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.6.0] - 2021-03-21
 
-[unreleased]: https://github.com/rogdham/bigxml/compare/v0.5.0...HEAD
+[0.6.0]: https://github.com/rogdham/bigxml/compare/v0.5.0...v0.6.0
+
+v0.6.0 adds documentation!
 
 ### :boom: Breaking changes